去除home模块使用drinkmodel，使用formula

+++ app/src/main/java/com/inno/coffee/function/formula/ProductProfileManager.kt
@@ -121,7 +121,8 @@ object ProductProfileManager {
 
         val componentProfileList =
             ComponentProfileList(componentList.size.toShort(), componentList)
-        return ProductProfile(formula.productId, preFlush, postFlush, componentProfileList)
+        return ProductProfile(formula.productId.toShort(), preFlush, postFlush,
+            componentProfileList)
     }
 
     suspend fun readFormulaFromAssets(context: Context) {
